app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
pydantic
passlib[bcrypt]
python-jose[cryptography]
mysql-connector-python
pydantic[email]

